// packages/shared/src/model/DTO/UserAssignmentDTO.ts
export interface UserAssignmentDTO {
    readonly alias: string
    readonly assignmentid: string
    readonly scoreearned: number
    readonly maxscore: number
    readonly gracedaysearned: number
    readonly gracedaysused: number
    readonly modifieddate: number
    readonly creditreceived: string
    readonly submitrange: string
}